components/tracking-page: Import slFlightTracking code
